Fix missing columns in CreateTables
I had been working on a version of CreateTables outside of my
git repository by accident while working on PopulateTables inside
the git repo. This resulted in PopulateTables trying to add stuff
to columns that don't exist, because only the version of CreateTables
that I was working on externaly created them.