Provide code to connect to database
There are a few variables inside the main function that can be
changed to specify the attributes of the server you are trying to
connect to.

The ones that are present are for connecting to the local server on my
laptop; these will have to be changed if you want to connect to the
server setup on your own computer. This will always be the case unless
we were able to set up a remote server (unlikely and probably not worth
the trouble)