Add checks for header being selected
Does not update selectedRow if header is selected as the header
index is -1 and causes crashes if the user attempts to expand
after clicking a header.
It is especially misleading because it leaves their previous
selection highlighted, given the impression that their selection
has not changed.
It now behaves as expected; the selectedRow does not update if
a header is clicked, meaning that the visual cue of a cell
being highlighted remains accurate